Leave the series naming to each interface's own get_metadata
The base demanded a name and a description through two hooks that raised, but
_add_time_series_segment_to_nwbfile already defaults name, description and unit
where the metadata does not state them, so the base was asserting a requirement
that is not one and forcing every subclass to answer it.

The hooks are gone and each interface fills metadata['TimeSeries'] itself, next
to the device entries and session start times it already states there. The base
keeps what was actually duplicated: the add_to_nwbfile body, where the series is
written, and the channel names.

The channel_ids property goes back to EDF, which was the only interface that
had it.
