+++ src/neuroconv/datainterfaces/ecephys/baserecordingtotimeseriesinterface.py
@@ -3,7 +3,6 @@ from typing import Literal
 from pynwb import NWBFile
 
 from ...basedatainterface import BaseDataInterface
-from ...utils import DeepDict
 
 
 class BaseRecordingToTimeSeriesInterface(BaseDataInterface):
@@ -15,9 +14,8 @@ class BaseRecordingToTimeSeriesInterface(BaseDataInterface):
     ``ElectricalSeries``, which is for electrical recordings from electrodes.
 
     A subclass reads its source into ``self.recording_extractor``, states ``self.metadata_key``, and
-    names the series through ``_get_time_series_name`` and ``_get_time_series_description``. A subclass
-    whose metadata carries more than the series, a session start time or a device, overrides
-    ``get_metadata`` and calls ``super()`` first.
+    names the series in its own ``get_metadata`` under ``metadata["TimeSeries"][self.metadata_key]``.
+    What it does not state, the writer defaults.
 
     A ``TimeSeries`` states one unit for all of its channels, so an interface here holds channels of one
     unit. Where a source mixes them, select one unit per interface rather than writing them together.
@@ -26,11 +24,6 @@ class BaseRecordingToTimeSeriesInterface(BaseDataInterface):
     # Where the series is written. A signal applied to the preparation is stimulus, a recorded one is
     # acquisition, and the writer defaults to acquisition.
     parent_container: Literal["acquisition", "stimulus"] = "acquisition"
-
-    @property
-    def channel_ids(self):
-        """The ids of the channels this interface holds, in the order they are written."""
-        return self.recording_extractor.get_channel_ids()
 
     def get_channel_names(self) -> list[str]:
         """
@@ -42,30 +35,6 @@ class BaseRecordingToTimeSeriesInterface(BaseDataInterface):
             The channel names, in the order they are written.
         """
         return list(self.recording_extractor.get_channel_ids())
-
-    def _get_time_series_name(self) -> str:
-        """The name of the written ``TimeSeries``. Subclasses state their own."""
-        raise NotImplementedError(
-            f"{type(self).__name__} must implement `_get_time_series_name` to name the TimeSeries it writes."
-        )
-
-    def _get_time_series_description(self) -> str:
-        """The description of the written ``TimeSeries``. Subclasses state their own."""
-        raise NotImplementedError(
-            f"{type(self).__name__} must implement `_get_time_series_description` to describe the TimeSeries it writes."
-        )
-
-    def get_metadata(self) -> DeepDict:
-        metadata = super().get_metadata()
-
-        # Assigned per key rather than replacing the block, so a subclass that filled it first keeps
-        # what it put there.
-        metadata["TimeSeries"][self.metadata_key] = dict(
-            name=self._get_time_series_name(),
-            description=self._get_time_series_description(),
-        )
-
-        return metadata
 
     def add_to_nwbfile(
         self,
